+++ collector/connect_godror.go
@@ -90,3 +90,20 @@ func isInvalidCredentialsError(err error) bool {
 	}
 	return oraErr.Code() == ora01017code || oraErr.Code() == ora28000code
 }
+
+func isTemporaryConnectionError(err error) bool {
+	err = errors.Unwrap(err)
+	if err == nil {
+		return false
+	}
+	oraErr, ok := err.(*godror.OraErr)
+	if !ok {
+		return false
+	}
+	switch oraErr.Code() {
+	case ora01033code, ora03113code, ora03114code, ora12537code:
+		return true
+	default:
+		return false
+	}
+}