+++ collector/connect_goora.go
@@ -80,3 +80,20 @@ func isInvalidCredentialsError(err error) bool {
 	}
 	return oraErr.ErrCode == ora01017code || oraErr.ErrCode == ora28000code
 }
+
+func isTemporaryConnectionError(err error) bool {
+	if err == nil {
+		return false
+	}
+	var oraErr *network.OracleError
+	ok := errors.As(err, &oraErr)
+	if !ok {
+		return false
+	}
+	switch oraErr.ErrCode {
+	case ora01033code, ora03113code, ora03114code, ora12537code:
+		return true
+	default:
+		return false
+	}
+}